plugin-connector: fold reconcileCursors back into connector-coordinator
reconcileCursors had exactly one caller (connector-coordinator's setCursors),
same as its single-target sibling createSingleCursor which already lives
inline in this module. Factoring it into util/ implied a general-purpose
helper it never was — move it (and its input types) next to
createSingleCursor, keep it exported so its existing unit test still reaches
it directly, and relocate that test to capabilities/ alongside the function.
