+++ packages/plugins/plugin-connector/src/capabilities/reconcile-cursors.test.ts
@@ -17,8 +17,8 @@ import { OperationInvoker } from '@dxos/operation';
 import { Expando } from '@dxos/schema';
 
 import { Connection, type ConnectorEntry, MaterializeTargetInput, MaterializeTargetOutput } from '../types';
-import { isCursorForConnection } from './cursor-predicates';
-import { type SyncTargetSelection, reconcileCursors } from './reconcile-cursors';
+import { isCursorForConnection } from '../util';
+import { type SyncTargetSelection, reconcileCursors } from './connector-coordinator';
 
 describe('reconcileCursors', () => {
   let builder: EchoTestBuilder;
