+++ packages/plugins/plugin-connector/src/util/index.ts
@@ -4,4 +4,3 @@
 
 export * from './connector-auth';
 export * from './cursor-predicates';
-export * from './reconcile-cursors';
